fix revalidation
