+++ apps/backend/src/publishHandler.ts
@@ -15,6 +15,7 @@ import type { z } from "zod";
 import type { CodemodVersionCreateInputSchema } from "../prisma/generated/zod";
 import type { CustomHandler } from "./customHandler";
 import { prisma } from "./db/prisma.js";
+import { buildRevalidateHelper } from "./revalidate";
 import { CLAIM_PUBLISHING } from "./services/tokenService.js";
 import { getCustomAccessToken } from "./util.js";
 
@@ -405,33 +406,8 @@ export const publishHandler: CustomHandler<Record<string, never>> = async ({
       }
     }
 
-    /**
-     * Revalidate codemod tag and page
-     */
-
-    const slug = buildCodemodSlug(name);
-
-    const revalidateURL = new URL(
-      `${environment.CODEMOD_COM_API_URL}/revalidate`,
-    );
-    revalidateURL.searchParams.set("path", `registry/${slug}`);
-
-    const revalidateTagURL = new URL(
-      `${environment.CODEMOD_COM_API_URL}/revalidate-tag`,
-    );
-    revalidateTagURL.searchParams.set("tag", `codemod-${slug}`);
-
-    try {
-      await fetch(revalidateTagURL.toString());
-      await fetch(revalidateURL.toString());
-
-      console.info(`Successfully revalidated ${slug}`);
-    } catch (e) {
-      const message = e instanceof Error ? e.message : String(e);
-      console.error(
-        `Failed to revalidate the page: ${slug}. Reason: ${message}`,
-      );
-    }
+    const revalidate = buildRevalidateHelper(environment);
+    await revalidate(name);
 
     return reply.code(200).send({ success: true });
   } catch (err) {
