+++ apps/frontend/app/(website)/registry/[codemod]/page.tsx
@@ -12,6 +12,8 @@ import { notFound } from "next/navigation";
 
 export const dynamicParams = true;
 
+const ONE_DAY = 60 * 60 * 24;
+
 export async function generateStaticParams() {
   const baseUrl = env.NEXT_PUBLIC_CODEMOD_AUTOMATIONS_LIST_ENDPOINT;
   const res = await fetchWithTimeout(`${baseUrl}/list`);
@@ -45,17 +47,20 @@ export async function generateMetadata(
 export default async function CodemodRoute({ params }) {
   const initialAutomationData = await loadCodemod(params.codemod, {
     next: {
-      revalidate: 60 * 60 * 24 * 30,
       tags: [`codemod-${params.codemod}`],
+      revalidate: ONE_DAY,
     },
   });
+
   if (!initialAutomationData || "error" in initialAutomationData) {
     notFound();
   }
 
   const automationPageData = await loadAutomationPage(
     initialAutomationData.tags,
+    ONE_DAY,
   );
+
   const pageData = transformAutomation({
     ...initialAutomationData,
     ...automationPageData?.data,
